+++ tests/TestCase/GlobalFunctionsTest.php
@@ -506,7 +506,11 @@ class GlobalFunctionsTest extends TestCase
      */
     public function testIsWinOnUnix()
     {
+        $this->assertIsDeprecated('is_win', 'The `is_win()` function is deprecated and will be removed in a later version. Use instead the `IS_WIN` constant');
+
+        $errorReporting = error_reporting(E_ALL & ~E_USER_DEPRECATED);
         $this->assertFalse(is_win());
+        error_reporting($errorReporting);
     }
 
     /**
@@ -516,7 +520,11 @@ class GlobalFunctionsTest extends TestCase
      */
     public function testIsWinOnWin()
     {
+        $this->assertIsDeprecated('is_win', 'The `is_win()` function is deprecated and will be removed in a later version. Use instead the `IS_WIN` constant');
+
+        $errorReporting = error_reporting(E_ALL & ~E_USER_DEPRECATED);
         $this->assertTrue(is_win());
+        error_reporting($errorReporting);
     }
 
     /**